Refactor MMA7455 sensor to individual module.
