+++ src/modules/editor.js
@@ -44,6 +44,7 @@ import './blockly/generators/propc/s3';
 import './blockly/generators/propc/sd_card';
 import './blockly/generators/propc/sensors';
 import './blockly/generators/propc/sensors/dht22';
+import './blockly/generators/propc/sensors/mma7455';
 import './blockly/generators/propc/sensors/mx2125';
 import './blockly/generators/propc/sensors/ping';
 import './blockly/generators/propc/sensors/pir';
